requirements.txt
rich
requests
python-dotenv